requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
Django==4.0.3
django-cleanup==6.0.0
django-filter==21.1
django-upgrade==1.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
msal==1.17.0
Pillow==9.0.1
promise==2.3
pycparser==2.21
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.1
requests==2.27.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tokenize-rt==4.2.1
urllib3==1.26.9
